Add comprehensive Google Cloud TPU and PyTorch/XLA support with multi-core training and audit

+++ maba/layers/gated_residual.py
@@ -7,7 +7,7 @@ class GatedRes(nn.Module):
         self.g_res = nn.Parameter(torch.full((dim,), bias_init, dtype=torch.float32))
 
     def forward(self, res: torch.Tensor, sub: torch.Tensor) -> torch.Tensor:
-        g = torch.sigmoid(self.g_res)
+        g = torch.sigmoid(self.g_res).type_as(res)
         return g * res + (1.0 - g) * sub
 
 GatedResidual = GatedRes
